[v0.91.4][maintainability] Split adl/src/cli/tests/pr_cmd_inline/mod.rs by responsibility (Closes #3432)
